+++ HISTORY.rst
@@ -24,6 +24,7 @@ latest (unreleased)
 * ALCYN-2232: Take into account OPW for cache issue on related fields on inheritsed model
 * ALCYN-2233: l10n_be_intrastat: propagate country_id
 * ALCYN-2167: Change the stock.location name computation, so a location as the same name if it is of usage 'view' or as the act_as_view flag set.
+* ALCYN:2246: Add a flag on stock location to mark the ones whose product quantity should not be used in the immediately available quantity computation
 
 **Bugfixes**